backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
aiofiles==23.2.1
